Interface/requirements.txt
streamlit
boto3
faiss-cpu
numpy
PyPDF2
python-docx
python-pptx
requests
sentencepiece
